MAGETWO-88054: Move declarative setup from setup to framework
