+++ app/code/Magento/CatalogSampleData/Setup/Patch/Data/InstallCatalogSampleData.php
@@ -7,8 +7,8 @@
 namespace Magento\CatalogSampleData\Setup\Patch\Data;
 
 use Magento\Framework\Setup;
-use Magento\Setup\Model\Patch\DataPatchInterface;
-use Magento\Setup\Model\Patch\PatchVersionInterface;
+use Magento\Framework\Setup\Patch\DataPatchInterface;
+use Magento\Framework\Setup\Patch\PatchVersionInterface;
 
 /**
  * Class InstallCatalogSampleData
