+++ app/code/Magento/ThemeSampleData/Setup/Patch/Data/InstallThemeSampleData.php
@@ -7,8 +7,8 @@
 namespace Magento\ThemeSampleData\Setup\Patch\Data;
 
 use Magento\Framework\Setup;
-use Magento\Setup\Model\Patch\DataPatchInterface;
-use Magento\Setup\Model\Patch\PatchVersionInterface;
+use Magento\Framework\Setup\Patch\DataPatchInterface;
+use Magento\Framework\Setup\Patch\PatchVersionInterface;
 
 /**
  * Class InstallThemeSampleData
